Wait for element to not be present.
Compatible with minor bok-choy upgrade, wherein 'not-present' and
'invisible' are two distinct concepts. In the bok-choy upgrade, an
invisible element must be present on the page. In the case of this
commit, the actual wait_for is expecting the element to disappear
from the page because it's injected by backbone while the page is
loading. Once it is removed, the page has loaded. So, the 'invisible'
distinction is not appropriate for what this test is doing. This
test was intended to search for not-present. Updating to make that so.